flask_server/requirements.txt
boto3
streamlit
PyPDF4
flask
textblob
flask-cors
chardet
requests
python-dotenv

